fix(conversion): three MSG/PST→EML fidelity fixes (round-16 audit)
Round-16 multi-angle audit of the MSG/PST/OST→EML conversion stack
(8 broad finders + 10 deep narrow finders, each finding adversarially
verified by a 3-lens panel). Three confirmed fidelity bugs, all in the
pst-parser/RTF layers; every fix has a regression test that fails on the
old behavior.

1. String8-typed PR_HTML decoded with the store code page, not the
   internet code page (Message/PropertyContext). A legacy PidTagBodyHtml
   stored as PtypString8 (id 0x1013) was re-decoded by the blanket
   decodeString8 pass with PR_MESSAGE_CODEPAGE, but the HTML body's bytes
   are governed by PR_INTERNET_CPID ([MS-OXCMAIL] §2.1.3.5.2) — the same
   rule the byte[]-typed variant and the MSG converter already follow. When
   the two code pages differ (e.g. a GBK store with a UTF-8 HTML body) the
   whole body mojibaked. decodeString8 now takes a per-tag charset override
   so PR_HTML alone decodes with the internet charset while every other
   String8 property keeps the store charset.

2. Named-property string-offset bounds check overflowed int and dropped
   the whole map (NameToIdMap). For a string-named NAMEID whose dwPropertyID
   (a String-stream byte offset read straight from disk) sat near
   Integer.MAX_VALUE, the guard `offset + 4 <= length` wrapped negative in
   32-bit arithmetic and passed, so position(offset) threw out of the
   unguarded parseStreams — discarding the store-wide named-property map
   (Keywords, all iCal appointment/VTODO props, …) for the entire store.
   The offset check is now widened to long, mirroring the already-correct
   length check on the next line.

3. RTF de-encapsulation leaked \bin binary payloads and {\pict}/{\object}
   groups into the recovered HTML (RtfStripper.deEncapsulateHtml and the
   PST fork Message.extractHtmlFromRtf). Unlike strip(), the de-encap loop
   neither consumed a \binN payload (RTF spec / [MS-OXRTFCP] §2.1.3.1.5)
   nor skipped picture/object groups, so their raw bytes leaked as body
   text and any stray brace in a payload popped the group-state stack and
   desynced all later output. Both forks now consume \binN wholesale (in
   the main loop and the group-skip helper) and drop {\pict}/{\object}
   whole, kept byte-identical between the two forks.

Audited but deliberately not changed: a PST From header can weld a
non-blank sender display-name onto the represented author's borrowed
address (PR_SENDER_NAME present, PR_SENDER_*_ADDRESS absent) where MSG
promotes the full author identity. The verification panel refuted it 2-1:
it is not a hard RFC 5322 §3.6.2 violation (the addr-spec belongs to the
author), the scenario is degenerate with no real fixture, and no clean
fix exists — a parser predicate breaks the delegate test doubles and a
driver-only fix introduces an opposite divergence for the genuine
same-address/different-name case.

+++ pst-parser/src/main/java/com/github/ttereshchenko/mailkit/pst/NameToIdMap.java
@@ -128,10 +128,13 @@ class NameToIdMap {
             if (isStringName) {
                 // The value is a signed offset into the string stream ([MS-PST] §2.4.7). A negative
                 // value would throw at position(...) and a value within 4 bytes of the end would
-                // underflow getInt(); validate before reading, and bound the payload with a long so a
-                // huge length cannot overflow the comparison. Node 0x61 is the store-wide NPID map, so a
-                // single malformed entry must not abort the whole map.
-                if (stringStream != null && propertyIdOrOffset >= 0 && propertyIdOrOffset + 4 <= stringStream.length) {
+                // underflow getInt(); validate before reading, and widen both the offset and the length
+                // bounds checks to long so a value near Integer.MAX_VALUE cannot overflow the comparison
+                // (offset + 4 wrapping negative) and slip a huge position(...) through. Node 0x61 is the
+                // store-wide NPID map, so a single malformed entry must not abort the whole map.
+                if (stringStream != null
+                        && propertyIdOrOffset >= 0
+                        && (long) propertyIdOrOffset + 4 <= stringStream.length) {
                     var stringBuffer = ByteBuffer.wrap(stringStream).order(ByteOrder.LITTLE_ENDIAN);
                     stringBuffer.position(propertyIdOrOffset);
                     int length = stringBuffer.getInt();
